chatbot-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
chromadb
sentence-transformers
google-generativeai
pydantic
numpy
python-multipart
